re-write to typescript
