--- .eslintrc
@@ -1,17 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "extends": "airbnb-base",
-  "env": {
-    "node": true,
-    "mocha": true
-  },
-  "rules": {
-    "func-names": 0,
-    "no-unused-vars": 0,
-    "global-require": 0,
-    "brace-style": 0,
-    "curly": 0,
-    "no-param-reassign": 0,
-    "no-else-return": 0
-  }
-}
