+++ package.json
@@ -4,19 +4,29 @@
   "description": "A json schema toolkit including functions for manipulating, traversing, and de-referencing json schema",
   "main": "build/index.js",
   "scripts": {
-    "clean": "rm -rf build",
-    "build": "npm test && npm run clean && babel src -d build",
-    "test": "mocha --compilers js:babel-core/register",
-    "prepublish": "npm run build"
+    "test": "jest --verbose",
+    "build": "./scripts/package",
+    "release": "cd dist && npm publish",
+    "lint": "tslint -p tsconfig.json --type-check",
+    "clean": "rm -rf dist",
+    "benchmark": "ts-node benchmark/index.ts"
   },
   "author": "Jason Wijegooneratne <admin@jwije.com>",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.2",
-    "babel-preset-es2015": "^6.18.0",
+    "@types/benchmark": "^1.0.30",
+    "@types/chai": "^3.5.2",
+    "@types/chalk": "^0.4.31",
+    "@types/jest": "^19.2.3",
+    "@types/node": "^7.0.21",
+    "benchmark": "^2.1.4",
     "chai": "^3.5.0",
-    "mocha": "^3.1.2"
+    "jest": "^20.0.4",
+    "ts-jest": "^20.0.4",
+    "ts-lint": "^4.5.1",
+    "ts-node": "^3.0.4",
+    "tslint-eslint-rules": "^4.1.0",
+    "typescript": "^2.3.3"
   },
   "dependencies": {
     "ajv": "^4.9.0",
