+++ scripts/package
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# Clean
+rm -rf dist
+mkdir dist
+
+# Copy latent files into ./dist
+cp * dist 2>&1 | grep -v 'omitting directory'
+cp .* dist 2>&1 | grep -v 'omitting directory'
+
+# Build typescript
+yarn tsc
+
+# Copy latent files from source, recursively
+cp -rp src/__tests__/schema dist/__tests__/schema
