+++ src/index.ts
@@ -1,13 +1,11 @@
-import Validator from './Validator';
 import { iterate, clone, contains, map, merge } from './object';
 import dereference from './dereference';
 
 export {
-  Validator,
-  clone,
-  contains,
-  dereference,
-  iterate,
-  map,
-  merge,
+    clone,
+    contains,
+    dereference,
+    iterate,
+    map,
+    merge,
 };
