+++ tslint.json
@@ -1,0 +1,22 @@
+{
+  "extends":[
+    "tslint:latest",
+    "tslint-eslint-rules"
+  ],
+  "rules": {
+    "quotemark": ["single"],
+    "only-arrow-functions": [false],
+    "curly": false,
+    "object-curly-spacing": [ true, "always" ],
+    "space-in-parens": [true, "never"],
+    "no-multi-spaces": [true],
+    "object-literal-sort-keys": false,
+    "indent": [
+      true, 2
+    ],
+    "member-access": false,
+    "interface-name": [
+      false, "never-prefix"
+    ]
+  }
+}
